Move service tests into test_init, remove test_service.py
